# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Nidria API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/cases": {
      "post": {
        "tags": [
          "cases"
        ],
        "summary": "Create Case",
        "operationId": "create_case_cases_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CaseCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "cases"
        ],
        "summary": "List Cases",
        "operationId": "list_cases_cases_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CaseStatus"
                  }
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "origin_country",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "pattern": "^[A-Z]{2}$"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Origin Country"
            }
          },
          {
            "name": "dest_country",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "pattern": "^[A-Z]{2}$"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Dest Country"
            }
          },
          {
            "name": "owner_agent_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Owner Agent Id"
            }
          },
          {
            "name": "preferred_lang",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Preferred Lang"
            }
          },
          {
            "name": "tag",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tag"
            }
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 25,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/cases/{case_id}/journey": {
      "post": {
        "tags": [
          "progress"
        ],
        "summary": "Assign Journey",
        "operationId": "assign_journey_cases__case_id__journey_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Case Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignJourneyRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StepProgressResponse"
                  },
                  "title": "Response Assign Journey Cases  Case Id  Journey Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/cases/{case_id}/steps": {
      "get": {
        "tags": [
          "progress"
        ],
        "summary": "Get Timeline",
        "operationId": "get_timeline_cases__case_id__steps_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Case Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StepProgressResponse"
                  },
                  "title": "Response Get Timeline Cases  Case Id  Steps Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/cases/{case_id}/steps/{step_progress_id}": {
      "patch": {
        "tags": [
          "progress"
        ],
        "summary": "Update Step",
        "operationId": "update_step_cases__case_id__steps__step_progress_id__patch",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Case Id"
            }
          },
          {
            "name": "step_progress_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Step Progress Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StepProgressUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StepProgressResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AssignJourneyRequest": {
        "properties": {
          "journey_template_id": {
            "type": "string",
            "format": "uuid",
            "title": "Journey Template Id"
          }
        },
        "type": "object",
        "required": [
          "journey_template_id"
        ],
        "title": "AssignJourneyRequest"
      },
      "BlockingStep": {
        "properties": {
          "template_step_id": {
            "type": "string",
            "format": "uuid",
            "title": "Template Step Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "template_step_id",
          "name"
        ],
        "title": "BlockingStep"
      },
...
      "CaseCreateRequest": {
        "properties": {
          "first_name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "First Name"
          },
          "last_name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Last Name"
          },
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "preferred_lang": {
            "type": "string",
            "maxLength": 5,
            "minLength": 2,
            "title": "Preferred Lang",
            "default": "fr"
          },
          "origin_country": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^[A-Z]{2}$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Origin Country"
          },
          "dest_country": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^[A-Z]{2}$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dest Country"
          },
          "status": {
            "$ref": "#/components/schemas/CaseStatus",
            "default": "prospect"
          },
          "source": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Source"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags"
          },
          "owner_agent_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner Agent Id"
          }
        },
        "type": "object",
        "required": [
          "first_name",
          "last_name",
          "email"
        ],
        "title": "CaseCreateRequest",
        "description": "Principal-only create: family members and external contacts go\nthrough their own endpoints."
      },
...
      "CaseListItemResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "agency_id": {
            "type": "string",
            "format": "uuid",
            "title": "Agency Id"
          },
          "principal_expat_user_id": {
            "type": "string",
            "format": "uuid",
            "title": "Principal Expat User Id"
          },
          "owner_agent_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner Agent Id"
          },
          "journey_template_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Journey Template Id"
          },
          "origin_country": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Origin Country"
          },
          "dest_country": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dest Country"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "principal": {
            "$ref": "#/components/schemas/PrincipalSummaryResponse"
          }
        },
        "type": "object",
        "required": [
          "id",
          "agency_id",
          "principal_expat_user_id",
          "owner_agent_id",
          "journey_template_id",
          "origin_country",
          "dest_country",
          "status",
          "source",
          "tags",
          "created_at",
          "updated_at",
          "principal"
        ],
        "title": "CaseListItemResponse"
      },
      "CaseListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/CaseListItemResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "page_size"
        ],
        "title": "CaseListResponse"
      },
...
      "CaseResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "agency_id": {
            "type": "string",
            "format": "uuid",
            "title": "Agency Id"
          },
          "principal_expat_user_id": {
            "type": "string",
            "format": "uuid",
            "title": "Principal Expat User Id"
          },
          "owner_agent_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner Agent Id"
          },
          "journey_template_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Journey Template Id"
          },
          "origin_country": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Origin Country"
          },
          "dest_country": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dest Country"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "agency_id",
          "principal_expat_user_id",
          "owner_agent_id",
          "journey_template_id",
          "origin_country",
          "dest_country",
          "status",
          "source",
          "tags",
          "created_at",
          "updated_at"
        ],
        "title": "CaseResponse"
      },
      "CaseStatus": {
        "type": "string",
        "enum": [
          "prospect",
          "in_progress",
          "awaiting_documents",
          "submitted",
          "validated",
          "closed"
        ],
        "title": "CaseStatus"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PrincipalSummaryResponse": {
        "properties": {
          "first_name": {
            "type": "string",
            "title": "First Name"
          },
          "last_name": {
            "type": "string",
            "title": "Last Name"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "preferred_lang": {
            "type": "string",
            "title": "Preferred Lang"
          }
        },
        "type": "object",
        "required": [
          "first_name",
          "last_name",
          "email",
          "preferred_lang"
        ],
        "title": "PrincipalSummaryResponse",
        "description": "Identity subset for list items — the full PrincipalResponse\n(activated) stays a detail-endpoint concern."
      },
...
      "ResponsibleType": {
        "type": "string",
        "enum": [
          "agent",
          "expat",
          "external"
        ],
        "title": "ResponsibleType",
        "description": "Polymorphic responsible of a step (also used for\n`journey_template_step.default_responsible_type`)."
      },
...
      "StepProgressResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "template_step_id": {
            "type": "string",
            "format": "uuid",
            "title": "Template Step Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "position": {
            "type": "integer",
            "title": "Position"
          },
          "estimated_days": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Estimated Days"
          },
          "required_documents": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Required Documents"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "responsible_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Responsible Type"
          },
          "responsible_agent_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Responsible Agent Id"
          },
          "responsible_external_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Responsible External Id"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "completed_by_agent_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed By Agent Id"
          },
          "blocked_by": {
            "items": {
              "$ref": "#/components/schemas/BlockingStep"
            },
            "type": "array",
            "title": "Blocked By"
          }
        },
        "type": "object",
        "required": [
          "id",
          "template_step_id",
          "name",
          "position",
          "estimated_days",
          "required_documents",
          "status",
          "responsible_type",
          "responsible_agent_id",
          "responsible_external_id",
          "completed_at",
          "completed_by_agent_id",
          "blocked_by"
        ],
        "title": "StepProgressResponse"
      },
      "StepProgressUpdateRequest": {
        "properties": {
          "status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/StepStatus"
              },
              {
                "type": "null"
              }
            ]
          },
          "responsible_type": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ResponsibleType"
              },
              {
                "type": "null"
              }
            ]
          },
          "responsible_agent_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Responsible Agent Id"
          },
          "responsible_external_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Responsible External Id"
          }
        },
        "type": "object",
        "title": "StepProgressUpdateRequest",
        "description": "Status transitions and/or responsible assignment. Unset fields\nare untouched (model_fields_set semantics); `responsible_type=None`\nexplicitly CLEARS the responsible."
      },
      "StepStatus": {
        "type": "string",
        "enum": [
          "todo",
          "in_progress",
          "blocked",
          "done"
        ],
        "title": "StepStatus"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/auth/agent/login"
          }
        }
      }
    }
  }
}